+++ Dxlib1/Game/Golem.cpp
@@ -64,7 +64,7 @@ void Golem::Update()
 	kY_ = indexY_;
 
 	// スピードによって移動のインターバルを変更
-	moveInterval_ = Field::CHIP_SIZE / speed_;
+	moveInterval_ = static_cast<int>(Field::CHIP_SIZE / speed_);
 
 	// 移動のインターバル
 	if (moveTimer_ % moveInterval_ == 0)
@@ -133,7 +133,7 @@ void Golem::Draw()
 
 		int imgY = DirectReturnNum(HEIGHT);
 
-		DrawRectRotaGraph(pos_.x, pos_.y - 5,		// 座標
+		DrawRectRotaGraph(static_cast<int>(pos_.x), static_cast<int>(pos_.y - 5),		// 座標
 			imgX, imgY,							// 切り取り左上
 			WIDTH, HEIGHT,						// 幅、高さ
 			SCALE, 0,							// 拡大率、回転角度
