+++ Dxlib1/Game/Slime.cpp
@@ -61,7 +61,7 @@ void Slime::Update()
 	kX_ = indexX_;
 	kY_ = indexY_;
 
-	moveInterval_ = Field::CHIP_SIZE / speed_;
+	moveInterval_ = static_cast<int>(Field::CHIP_SIZE / speed_);
 
 	// 移動のインターバル
 	if (moveTimer_ % moveInterval_ == 0)
@@ -130,7 +130,7 @@ void Slime::Draw()
 
 		int imgY = DirectReturnNum(HEIGHT);
 
-		DrawRectRotaGraph(pos_.x, pos_.y,		// 座標
+		DrawRectRotaGraph(static_cast<int>(pos_.x), static_cast<int>(pos_.y),		// 座標
 			imgX, imgY,							// 切り取り左上
 			WIDTH, HEIGHT,						// 幅、高さ
 			SCALE, 0,							// 拡大率、回転角度
