+++ Dxlib1/Scene/SceneManager.cpp
@@ -34,7 +34,7 @@ void SceneManager::Update(const InputState& input)
 
 void SceneManager::Draw()
 {
-	for (int i = scene_.size() - 1; i >= 0; --i)
+	for (int i = static_cast<int>(scene_.size()) - 1; i >= 0; --i)
 	{
 		scene_[i]->Draw();
 	}
